Interconnectivity Top Bar Update
Changed the background of the top bar of the Interconnectivity Page to be the image used on the home page as I believe the effect to be quite aesthetically pleasing. I will look into ensuring that the image is positioned in the same way as on the main screen, and am seeking approval to mirror this change across all bootstrap changes; if this is not acceptable, please tell me and I will revert the change for the sake of conformity.