Healthcare Page Top Bar Style
Linked the Standard Bootstrap Stylesheet to the Healthcare page and edited the classes of the elements in its top bar so as to apply the standard top bar styling to the page in question. Also took the opportunity whilst editing the page to add a missing link to the Communication Page to the nav bar.