Communication Page Top Bar Style
Applied the Standard Bootstrap styling to the Top Bar of the Communication Page. With this, the Interconnectivity, Healthcare and Communication Pages all have the same top bars and nav bars so mesh together very nicely and fit well with the main page, however the Research and Entertainment pages remain different as they do not currently implement Bootstrap and hence cannot easily import the Bootstrap top bar present in the afoermentioned three pages.